Reduce code and added comments

+++ view/npo_view.py
@@ -72,15 +72,6 @@ class NpoWindow(Screen):
         else:
             self.ids.name_labels.text = "No NPO selected"
 
-    # Not in use
-    def press_all_npos(self):
-        n_list = npoController().get_all_npos()
-        text = "The NPOs are:"
-        for i in n_list:
-            text += f'{i["name"]}, '
-
-        self.ids.name_labels.text = text
-
 
 # Update Window
 class NpoUpdateWindow(Screen):
@@ -119,41 +110,7 @@ class NpoCreateWindow(Screen):
             if npo_exists["n_id"] == -1:
                 npo_exists["name"] = name
 
-            NpoCreatePop(npo_exists).open_pop()
-
-
-# Create Popup Window
-class NpoCreatePop(FloatLayout):
-    def __init__(self, npo):
-        super().__init__()
-        self.npo_id = npo["n_id"]
-        self.npo_name = npo["name"]
-
-    # Create popup with message
-    def open_pop(self):
-        self.message(self.npo_name)
-        self.popup = Popup(title="PopupWindow", content=self, size_hint=(None, None), size=(700, 700))
-        self.popup.open()
-
-    def close_pop(self):
-        self.popup.dismiss()
-
-    def message(self, name):
-        # If no NPO, add message to confirm
-        if self.npo_id == -1:
-            self.ids.name_label.text = f'Are you sure you want to create the NPO {name}?'
-            self.ids.button_name.text = "Confirm"
-
-        # If a NPO exist, throw exist message
-        else:
-            self.ids.name_label.text = f'NPO {name} already exist, cannot create'
-            self.npo_name = "None"
-            self.ids.button_name.text = "Return"
-
-    def create(self, name):
-        dict = {"name": name}
-        create = npoController().create_npo(dict)
-        return create
+            NpoPop(npo_exists, "Create").open_pop()
 
 
 # Delete Window
@@ -161,6 +118,7 @@ class NpoDeleteWindow(Screen):
 
     def on_pre_enter(self, *args):
         self.ids.NPO_dropdown.text = "NPOs"
+        self.ids.NPO_dropdown.values = self.manager.get_screen("first").ids.NPO_dropdown.values
 
     # Delete NPO entry in the database. Open Popup to confirm value.
     def remove_npo(self):
@@ -170,16 +128,16 @@ class NpoDeleteWindow(Screen):
             if npo_exists["n_id"] == -1:
                 npo_exists["name"] = name
 
-            NpoDeletePop(npo_exists).open_pop()
+            NpoPop(npo_exists, "Delete").open_pop()
 
 
-# Delete Window
-class NpoDeletePop(Screen):
-    # Delete NPO entry in the database. Open Popup to confirm value.
-    def __init__(self, npo):
+# Create Popup Window
+class NpoPop(FloatLayout):
+    def __init__(self, npo, action):
         super().__init__()
         self.npo_id = npo["n_id"]
         self.npo_name = npo["name"]
+        self.action = action
 
     # Create popup with message
     def open_pop(self):
@@ -191,18 +149,40 @@ class NpoDeletePop(Screen):
         self.popup.dismiss()
 
     def message(self, name):
-        # If a NPO exist, add message to confirm delete
-        if self.npo_id != -1:
-            self.ids.name_label.text = f'Are you sure you want to delete the NPO {name}?'
-            self.ids.button_name.text = "Confirm"
+        # Create Messages
+        if self.action == "Create":
+            # If no NPO, add message to confirm
+            if self.npo_id == -1:
+                self.ids.name_label.text = f'Are you sure you want to create the NPO {name}?'
+                self.ids.button_name.text = "Confirm"
 
-        # If a NPO does not exist, throw not found message
-        else:
-            self.ids.name_label.text = f'NPO {name} was not found'
-            self.npo_name = "None"
-            self.ids.button_name.text = "Return"
+            # If a NPO exist, throw exist message
+            else:
+                self.ids.name_label.text = f'NPO {name} already exist, cannot create'
+                self.npo_name = "None"
+                self.ids.button_name.text = "Return"
 
-    def delete(self, name):
-        dict = {"name": name}
-        delete = npoController().delete_npo(dict)
-        return delete
+        # Delete Messages
+        elif self.action == "Delete":
+            # If a NPO exist, add message to confirm delete
+            if self.npo_id != -1:
+                self.ids.name_label.text = f'Are you sure you want to delete the NPO {name}?'
+                self.ids.button_name.text = "Confirm"
+
+            # If a NPO does not exist, throw not found message
+            else:
+                self.ids.name_label.text = f'NPO {name} was not found'
+                self.npo_name = "None"
+                self.ids.button_name.text = "Return"
+
+# Crud operations for the Pop Up window
+    def crud_action(self, name):
+        if self.action == "Create":
+            dict = {"name": name}
+            create = npoController().create_npo(dict)
+            return create
+
+        elif self.action == "Delete":
+            dict = {"name": name}
+            delete = npoController().delete_npo(dict)
+            return delete
